[Samples] Fixed several warnings. Fixed WYZ sample data. [Trilo SCC] Fixed undocumented instruction. [Build tool] Clean code. Added support for NEO-8/16 mapper format. [Doc] Updated source code documentation. Fixed several typo issues.

+++ engine/src/scroll.c
@@ -159,7 +159,7 @@ void Scroll_HBlankAdjust(u8 adjust)
 		break;
 	case 1:
 		VDP_SetAdjustOffset(g_Scroll_Adjust);
-		VDP_SetHBlankLine((SCROLL_DST_Y + SCROLL_DST_H) * 8 - 4);
+		VDP_SetHBlankLine((u8)((SCROLL_DST_Y + SCROLL_DST_H) * 8 - 4));
 		break;
 	default:
 		VDP_SetAdjustOffset(0);
